feat(keyboard): add keyboard layout configuration for ANSI QWERTY, move the 'type_def.py' to the 'type' folder
